feat: split change group easier to test, read, and update
fix: add blocked users to created group and updated group

fix: update dynamodb config [UNTESTED]
